+++ python/lsst/ts/mtdome/config_schema.py
@@ -26,8 +26,8 @@ import yaml
 CONFIG_SCHEMA = yaml.safe_load(
     """
     $schema: http://json-schema.org/draft-07/schema#
-    $id: https://github.com/lsst-ts/ts_mtdome/blob/master/python/lsst/ts/mtdome/config_schema.py
-    title: MTDome v1
+    $id: https://github.com/lsst-ts/ts_mtdome/blob/main/python/lsst/ts/mtdome/config_schema.py
+    title: MTDome v2
     description: Schema for MTDome configuration files
     type: object
     properties:
@@ -35,21 +35,17 @@ CONFIG_SCHEMA = yaml.safe_load(
         description: IP address of the TCP/IP interface
         type: string
         format: hostname
-        default: "127.0.0.1"
       port:
         description: Port number of the TCP/IP interface
         type: integer
-        default: 5000
       connection_timeout:
         description: Time limit for connecting to the TCP/IP interface (sec)
         type: number
         exclusiveMinimum: 0
-        default: 10
       read_timeout:
         description: Time limit for reading data from the TCP/IP interface (sec)
         type: number
         exclusiveMinimum: 0
-        default: 10
     required:
       - host
       - port